feat: add Collapsible and Dropdown components; enhance EventHistory and Events pages with new functionality and UI improvements
Collapsible is the new component for all visualizations on a team, event, or player page, (old code still needs to be updated, like eventhistory.tsx)

Dropdown is an update to the filter system to allow better management and reusability via component

Searchable dropdown was updated to add arrow functionality and now scrolls alongside arrows (same functionality still needs to be applied to dropdown)

EventHistory was updated to support interconnectedness between pages (dom navigation). The event cards in the eventhistory timeline are now clickable and direct you to the respective events pages on ValorantAnalyticEngine website.

Events.tsx was updated to align more with the design of other pages + added a placements table using a newly added rpc call to the supabase database (resultsbyevent). It uses the new Collapsible and implements the updated searchabledropdown for event selection and dropdown for year selection filter. Also made link to liquipedia a part of the event name and linked to external liquipedia team names pages for the placements table in the collapsible.

index.ts was updated with the eventplacement interface for the new supabase rpc function call

sqlfunctions was updated with the new function call and also updated existing function 'eventplacementsbyorgid' to return event_id instead of event_org_id for better generalization and to enable connectivity from event history timeline to events page